Was able to get MapsActivity to work. I think I've enable everything with respect to the Google Maps API so we can use what we think would be google or useful.
Note: I updated the git path to work with the git.exe on my computer which may cause an error in yours. If it is a problem for you, just update the path to your unique one.

Also, I've secured the API key so that it only work with Android.  If we need it for something else, please let me know.